// src/app/shared/models/ventas/facturacion/IPaymentItem.ts
export interface IPaymentItem {
    nVenPayCodigo: number;
    nVenHeaCodigo: number;
    nVenDetItem: number;
    nVenDetArtMasCodigo: number;
    nVenDetMoneda: number;
    nVenDetImportePago: number;
    nVenPayTC: number;
    nVenPayEquivalente: number;
    nVenDetTipLinea: number;
    dVenPayFecRegistro: string;
    cVenPayUsuRegistro: string;
    cVenPayHostRegistro: string;
    dVenPayFecUpdate: string;
    cVenPayUsuUpdate: string;
    cVenAccDetHostUpdate: string;
}